requirements.txt
streamlit==1.28.0
sentence-transformers==2.2.2
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
PyPDF2==3.0.1
python-docx==0.8.11
plotly==5.15.0
nltk==3.8.1
spacy==3.6.1
faiss-cpu==1.7.4 